# Use an interval tree for SIR labels.