--- a/yktrace/Cargo.toml
+++ b/yktrace/Cargo.toml
@@ -11,6 +11,7 @@ fallible-iterator = "0.2.0"
 fxhash = "0.2.1"
 gimli = "0.23.0"
 hwtracer = { git = "https://github.com/softdevteam/hwtracer" }
+intervaltree = "0.2"
 lazy_static = "1.4.0"
 libc = "0.2.80"
 memmap = "0.7.0"