To make a SIR trace, there's a "mapping" stage which translates PT virtual addresses to SIR locations. Before this change, for each block libipt gave us we had to do a linear search over a list of label addresses, searching for those that coincide. That was slow.

This change puts the labels in a data structure better suited: an interval tree. Then computing the coinciding intervals can be done in log time, rather than linear time.

Rough benchmarking of test suite:

before:
```
1: cargo test
            Mean        Std.Dev.    Min         Median      Max
real        9.363       0.104       9.237       9.395       9.480
user        98.112      0.361       97.681      98.081      98.720
sys         6.037       0.553       5.157       6.012       6.804
```

after:
```
            Mean        Std.Dev.    Min         Median      Max
real        4.574       0.177       4.310       4.543       4.794
user        19.170      0.180       18.889      19.195      19.382
sys         9.995       0.273       9.666       9.997       10.446
```